requirements.txt
langchain
langchain_core
langchain-groq
langchain-openai
langchain-anthropic
pandas
streamlit
python-dotenv